+++ client/Assets/Valgor/City/UI/BuildingContextAction.cs
@@ -14,7 +14,8 @@ namespace Valgor.City.UI
         Train,
         Research,
         Open,
-        Send
+        Send,
+        Feed
     }
 
     public readonly struct BuildingContextActionInfo